fix comments in the second review
