+++ components/log-viewer-webui/client/src/components/ResultsTimeline/datetime.ts
@@ -27,85 +27,6 @@ enum TIME_UNIT {
     YEAR = "year",
 }
 
-enum TIME_RANGE_MODIFIER {
-    NONE = "none",
-    TODAY = "today",
-    LAST = "last",
-    PREV = "prev",
-    TO_DATE = "to-date",
-}
-
-/**
- * Time range presets.
- *
- * @type {{[key: string]: string}}
- */
-const TIME_RANGE_PRESET_LABEL = Object.freeze({
-    [`${TIME_UNIT.MINUTE}_${TIME_RANGE_MODIFIER.LAST}_15`]: "Last 15 Minutes",
-    [`${TIME_UNIT.MINUTE}_${TIME_RANGE_MODIFIER.LAST}_60`]: "Last 60 Minutes",
-    [`${TIME_UNIT.HOUR}_${TIME_RANGE_MODIFIER.LAST}_4`]: "Last 4 Hours",
-    [`${TIME_UNIT.HOUR}_${TIME_RANGE_MODIFIER.LAST}_24`]: "Last 24 Hours",
-    [`${TIME_UNIT.DAY}_${TIME_RANGE_MODIFIER.PREV}_1`]: "Previous Day",
-    [`${TIME_UNIT.WEEK}_${TIME_RANGE_MODIFIER.PREV}_1`]: "Previous Week",
-    [`${TIME_UNIT.MONTH}_${TIME_RANGE_MODIFIER.PREV}_1`]: "Previous Month",
-    [`${TIME_UNIT.YEAR}_${TIME_RANGE_MODIFIER.PREV}_1`]: "Previous Year",
-    [`${TIME_UNIT.DAY}_${TIME_RANGE_MODIFIER.TODAY}_0`]: "Today",
-    [`${TIME_UNIT.WEEK}_${TIME_RANGE_MODIFIER.TO_DATE}_0`]: "Week to Date",
-    [`${TIME_UNIT.MONTH}_${TIME_RANGE_MODIFIER.TO_DATE}_0`]: "Month to Date",
-    [`${TIME_UNIT.YEAR}_${TIME_RANGE_MODIFIER.TO_DATE}_0`]: "Year to Date",
-    [`${TIME_UNIT.ALL}_${TIME_RANGE_MODIFIER.NONE}_0`]: "All Time",
-});
-
-/**
- * Computes a time range based on a token.
- *
- * @param token representing the time range to compute; format: `unit_modifier_amount`
- * @return The computed time range
- */
-const computeTimeRange = (token: string): TimeRange => {
-    const [
-        unit,
-        modifier,
-        amount,
-    ] = token.split("_") as [TIME_UNIT, TIME_RANGE_MODIFIER, number];
-    let end;
-    let begin;
-
-    if (TIME_UNIT.ALL === unit) {
-        end = dayjs.utc().add(1, TIME_UNIT.YEAR);
-        begin = dayjs.utc(0);
-    } else {
-        const isEndingNow = [
-            TIME_RANGE_MODIFIER.LAST,
-            TIME_RANGE_MODIFIER.TODAY,
-            TIME_RANGE_MODIFIER.TO_DATE,
-        ].includes(modifier);
-        const isBeginStartOfUnit = [
-            TIME_RANGE_MODIFIER.PREV,
-            TIME_RANGE_MODIFIER.TODAY,
-            TIME_RANGE_MODIFIER.TO_DATE,
-        ].includes(modifier);
-
-        end = (true === isEndingNow) ?
-            dayjs.utc() :
-            dayjs.utc()
-                .subtract(amount, unit)
-                .endOf(unit);
-        begin = (true === isBeginStartOfUnit) ?
-            end.startOf(unit) :
-            end.subtract(amount, unit);
-    }
-
-    return {
-        begin,
-        end,
-    };
-};
-
-const DEFAULT_TIME_RANGE: TimeRange = computeTimeRange(
-    `${TIME_UNIT.ALL}_${TIME_RANGE_MODIFIER.NONE}_0`,
-);
-
 // eslint-disable-next-line no-warning-comments
 // TODO: Switch date pickers so we don't have to do this hack
 /**
@@ -180,13 +101,10 @@ const convertZoomTimestampToUtcDatetime = (timestampUnixMillis: number) => {
 
 export type {TimeRange};
 export {
-    computeTimeRange,
     convertLocalDateToSameUtcDatetime,
     convertUtcDatetimeToSameLocalDate,
     convertZoomTimestampToUtcDatetime,
     DATETIME_FORMAT_TEMPLATE,
-    DEFAULT_TIME_RANGE,
     expandTimeRangeToDurationMultiple,
-    TIME_RANGE_PRESET_LABEL,
     TIME_UNIT,
 };
