+++ components/log-viewer-webui/client/src/components/ResultsTimeline/index.module.css
@@ -2,6 +2,5 @@
   cursor: var(--timeline-chart-cursor, crosshair);
   height: 150px;
   max-height: 150px;
-  border-top: 1px solid #aeaeae;
   border-bottom: 1px solid #aeaeae;
 }
